refactor: Prize domain.prize로 패키지 이동
