+++ src/main/java/ladder/domain/prize/Prize.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.prize;
 
 public class Prize {
